chore: clippy warnings

+++ cli/src/display.rs
@@ -89,7 +89,7 @@ impl<U: DisplayUserFacing<C>, C: Color> Display for FmtUserFacing<'_, U, C> {
 
         let nr_width = visible_lines
             .last()
-            .map_or(2, |hl_line| (hl_line.nr as f32).log10() as usize + 1);
+            .map_or(2, |hl_line| (hl_line.nr as f64).log10() as usize + 1);
 
         if let Some((first, others)) = visible_lines.split_first() {
             let hl_line = first;
